Moved the search component to modals
